finish: details and page header changes
