--- lib/components/Details/_styles.tsx
@@ -1,11 +1,0 @@
-export const listStyles = {
-  listWrapper: `cu-pageheader-group mb-6 mt-6`,
-  listGroup: `pl-0 list-none`,
-  listHorizontal: `flex flex-wrap gap-6 items-center`,
-  listLink: 'font-semibold text-cu-red hover:underline',
-}
-
-export const socialStyles = {
-  link: 'text-cu-black-400',
-  icon: 'w-6 h-6 md:w-8 md:h-8',
-}
